add ics cleanup config option
Some ics files come with invalid formatting that the ics library does
not allow. This commit adds a config option to define one or more regex
rules for replacing potantially breaking parts of the ics.

Also add example usage with such a broken feed in the
config.json.example file.

Signed-off-by: Nils Herde <nils@yvn.no>